% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{binding_frames_mm}
\alias{binding_frames_mm}
\title{binding_frames_mm}
\usage{
binding_frames_mm(x, auto.corrected, add.col)
}
\arguments{
\item{x}{Wesame-style data frame}

\item{auto.corrected}{Whether or not data went through AutoCorrecPeaks}

\item{add.col}{Addtional (numeric) column to add to output dataframe to be
used in downstream analysis/plotting}
}
\value{
Segment dataframe output of multiple dfs bound together
}
\description{
Bind individual sesame frames together
}
