% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{calc_seg_state}
\alias{calc_seg_state}
\title{Bind individual sesame frames together}
\usage{
calc_seg_state(seg, cutoff)
}
\arguments{
\item{cutoff}{cutoff CN, any number above this will be assigned this value}

\item{x}{sesame-style data frame}
}
\value{
seg frame with state assigned
}
\description{
Bind individual sesame frames together
}
