% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.EpiMod}
\alias{champ.EpiMod}
\title{champ.EpiMod}
\usage{
champ.EpiMod(
  beta = myNorm,
  pheno = myLoad$pd$Sample_Group,
  nseeds = 100,
  gamma = 0.5,
  nMC = 1000,
  sizeR.v = c(1, 100),
  minsizeOUT = 10,
  resultsDir = "./CHAMP_EpiMod/",
  PDFplot = TRUE,
  arraytype = "450K"
)
}
\arguments{
\item{beta}{The beta parameter}

\item{pheno}{The pheno parameter}

\item{nseeds}{The nseeds parameter}

\item{gamma}{The gamma parameter}

\item{nMC}{The nMC parameter}

\item{sizeR.v}{The sizeR.v parameter}

\item{minsizeOUT}{The minisizeOUT parameter}

\item{resultsDir}{The resultsDir parameter}

\item{PDFplot}{The PDFplot parameter}

\item{arraytype}{The arraytype parameter}
}
\value{
A ChAMP EpiMod object
}
\description{
MethylMaster version of the champ.EpiMod() function
Note: BiocManager::install("FEM")
see https://rdrr.io/github/gaberosser/ChAMP/src/R/champ.EpiMod.R
}
