% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.import.mm}
\alias{champ.import.mm}
\title{champ.import.mm}
\usage{
champ.import.mm(
  directory = getwd(),
  sample.sheet = NULL,
  offset = 100,
  control = TRUE,
  compare.group = compare.group,
  arraytype = "450K"
)
}
\arguments{
\item{directory}{The input directory}

\item{sample.sheet}{The MethylMaster sample sheet}

\item{offset}{The offset to use}

\item{control}{Whether or not to use control}

\item{compare.group}{The MethylMaster two-element comparison vector to use}

\item{arraytype}{The array type eg. "450k"}

\item{infile}{The path to the input file}
}
\value{
Returns a list of imported data
}
\description{
MethylMaster modified champ.import() function
}
