% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.load.mm}
\alias{champ.load.mm}
\title{champ.load.mm}
\usage{
champ.load.mm(
  directory = getwd(),
  sample.sheet = NULL,
  compare.group = compare.group,
  control = FALSE,
  method = "ChAMP",
  methValue = "B",
  autoimpute = TRUE,
  filterDetP = TRUE,
  ProbeCutoff = 0,
  SampleCutoff = 0.1,
  detPcut = 0.01,
  filterBeads = TRUE,
  beadCutoff = 0.05,
  filterNoCG = TRUE,
  filterSNPs = TRUE,
  population = NULL,
  filterMultiHit = TRUE,
  filterXY = TRUE,
  force = FALSE,
  arraytype = "450K"
)
}
\value{
Return a list of loaded data
}
\description{
champ.load function modified by Michael Mariani
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kept.
}
