% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{cnvOncoPrint}
\alias{cnvOncoPrint}
\title{cnvOncoPrint}
\usage{
cnvOncoPrint(
  calls,
  features,
  multi.calls = .largest,
  top.features = 25,
  top.samples = 100,
  ...
)
}
\arguments{
\item{calls}{The calls parameter}

\item{features}{The features parameter}

\item{multi.calls}{The multi.calls parameter}

\item{top.features}{The top.features parameter}

\item{top.samples}{The top.samples parameter}

\item{...}{Additional parameters passed to cnvOncoPrint}
}
\description{
The MethylMaster version of the OncoPrint() function
to:
"plot for CNV regions
Illustrates overlaps between CNV calls and genomic features across a
sample population."
}
