% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{DoEpiMod}
\alias{DoEpiMod}
\title{DoEpiMod}
\usage{
DoEpiMod(
  intEpi.o,
  nseeds = 100,
  gamma = 0.5,
  nMC = 1000,
  sizeR.v = c(1, 100),
  minsizeOUT = 10,
  writeOUT = TRUE,
  nameSTUDY = "X",
  ew.v = NULL
)
}
\arguments{
\item{intEpi.o}{The intEpi.o parameter}

\item{nseeds}{The nseeds parameter}

\item{gamma}{The gamma parameter}

\item{sizeR.v}{The sizeR.v parameter}

\item{minsizeOUT}{The minsizeout parameter}

\item{writeOUT}{The writeOUT parameter}

\item{nameSTUDY}{The nameSTUDY parameter}

\item{ew.v}{The ew.v parameter}

\item{nMC=1000}{The nMC parameter}
}
\value{
A list of results
}
\description{
MethylMaster version of the DoEpiMod function
https://rdrr.io/bioc/FEM/src/R/DoEpiMod.R
}
