% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{DoIntEpi450k}
\alias{DoIntEpi450k}
\title{DoIntEpi450k}
\usage{
DoIntEpi450k(statM.o, adj.m, c, dmaMode = "avbeta")
}
\arguments{
\item{statM.o}{The statM.o parameter}

\item{adj.m}{The adj.m parameter}

\item{c}{The c parameter}

\item{dmaMode}{The dmaMode parameter}
}
\value{
A list of results
}
\description{
The MethylMaster version of the DoIntEpi450k function
See https://rdrr.io/bioc/FEM/src/R/DoIntEpi450k.R
}
