% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.clusterCalls}
\alias{.clusterCalls}
\title{.clusterCalls}
\usage{
.clusterCalls(calls, ro.thresh = 0.5, multi.assign = FALSE)
}
\arguments{
\item{calls}{GenomicRanges::GRanges}

\item{ro.thresh}{numeric}

\item{multi.assign}{logical}
}
\value{
list of integer vectors
}
\description{
MethylMaster version of the .clusterCalls function
"the clustering itself then goes sequentially through the identified RO hits,
touching each hit once, and checks whether this hit could be merged to
already existing clusters"
}
