% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{.coerce.pData.mm}
\alias{.coerce.pData.mm}
\title{.coerce.pData.mm}
\usage{
.coerce.pData.mm(pdat)
}
\arguments{
\item{pdat}{Path to the input file}
}
\value{
compatible pData data frame
}
\description{
This will fix the issue with LRR2
}
