% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{.funnorm.mm}
\alias{.funnorm.mm}
\title{.funnorm.mm}
\usage{
.funnorm.mm(rgSet, nPCs = 2, sex = NULL, verbose = TRUE)
}
\arguments{
\item{rgSet}{The rgSet parameter}

\item{nPCs}{The nPCs parameter}

\item{sex}{The sex parameter}

\item{verbose}{the verbose parameter}
}
\value{
Return a final gm set
}
\description{
MethylMaster version of the Epicopy .funnorm() function
}
