% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.getFreq.mm}
\alias{.getFreq.mm}
\title{.getfreq.mm}
\usage{
.getFreq.mm(calls, type = c("gain", "loss"))
}
\arguments{
\item{calls}{GenomicRanges::GRanges}

\item{type}{character}
}
\value{
numeric
}
\description{
MethylMaster version of the .getFreq() function
}
