% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.getROHits}
\alias{.getROHits}
\title{.getROHits}
\usage{
.getROHits(calls, ro.thresh = 0.5)
}
\arguments{
\item{calls}{GenomicRanges::GRanges}

\item{ro.thresh}{numeric}
}
\value{
S4Vectors::Hits
}
\description{
.getROHits
}
