% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.getType.mm}
\alias{.getType.mm}
\title{.getType.mm}
\usage{
.getType.mm(states, type.thresh = 0.1)
}
\arguments{
\item{states}{integer vector}

\item{state.thresh}{numeric}
}
\value{
character
}
\description{
The MethylMaster version of .getType() function
}
