% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.pruneMultiAssign}
\alias{.pruneMultiAssign}
\title{.pruneMultiAssign}
\usage{
.pruneMultiAssign(clusters)
}
\arguments{
\item{clusters}{list of integer vectors}
}
\value{
list of integer vectors
}
\description{
the MethylMaster version of the .pruneMultiAssign function
}
