% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.roPopRanges}
\alias{.roPopRanges}
\title{.roPopRanges}
\usage{
.roPopRanges(grl, ro.thresh = 0.5, multi.assign = FALSE, verbose = FALSE)
}
\arguments{
\item{grl}{The grl parameter}

\item{multi.assign}{The multu.assign parameter}

\item{verbose}{The verbose parameter}

\item{ro.tresh}{The ro.thresh parameter}
}
\value{
GenomicRanges::GRanges
}
\description{
MethylMaster version of the .roPopRanges function
"(2) RO approach"
}
