% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{.scoreRegion}
\alias{.scoreRegion}
\title{.scoreRegion}
\usage{
.scoreRegion(r, calls, fract.thresh = 0.1)
}
\arguments{
\item{r}{GenomicRanges::GRanges}

\item{calls}{GenomicRanges::GRanges}

\item{fract.thresh}{numeric}
}
\value{
numeric
}
\description{
The MethylMasteR version of the .scoreRegion function
}
