% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{epicopy.mm}
\alias{epicopy.mm}
\title{Modified epicopy function from Epicopy}
\usage{
epicopy.mm(
  target_dir,
  output_dir = NULL,
  single.file = TRUE,
  single.file.path = NULL,
  comparisons = NULL,
  reference_group = "normal",
  project_name = NULL,
  Normals = NULL,
  sampNames = NULL,
  QN = FALSE,
  Ref = "mode",
  mode.bw = 0.1,
  mode.method = "naive",
  normal.cnv = NULL,
  mean.center = TRUE,
  filterProbes = FALSE,
  retainedProbes = NULL,
  ncores = 1L,
  seg.undo.splits = "sdundo",
  seg.undo.SD = 2,
  filterbycount = TRUE,
  min_probes = 50,
  verbose = TRUE,
  keep_fnobj = FALSE,
  fn_output = NULL,
  ...
)
}
\value{
return a CNV frame
}
\description{
I updated the read.metharray.sheet function
and incorporated it here
}
