% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{fast_450k_binning}
\alias{fast_450k_binning}
\title{This is a faster version of binding 450K bins}
\usage{
fast_450k_binning(
  proc.samples,
  med.ref.values,
  proc.ref.samples,
  bin.size = 5e+05
)
}
\arguments{
\item{proc.samples}{samples input}

\item{med.ref.values}{median reference values}

\item{proc.ref.samples}{refernce samples}

\item{bin.size}{bin size}
}
\value{
450k-style candidates data frame
}
\description{
This is a faster version of binding 450K bins
}
