% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{FemModShow}
\alias{FemModShow}
\title{FemModShow}
\usage{
FemModShow(mod, name, fem.o, mode = "Integration")
}
\arguments{
\item{mod}{The mod parameter}

\item{name}{The name parameter}

\item{fem.o}{The fem.o parameter}

\item{mode}{The mode parameter}
}
\value{
Perform caluclations and plotting
}
\description{
MethylMaster verssion of the FemModShow() function
See https://rdrr.io/bioc/FEM/src/R/FemModShow.R
}
