% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_findSegments2.R
\name{findSegments2}
\alias{findSegments2}
\title{findSegments2}
\usage{
findSegments2(
  data,
  ctrl,
  ctrlAll,
  statistic = "wilcoxon",
  plot = FALSE,
  delta = 500,
  output = "diff",
  ylim = NULL,
  arrayType = "auto"
)
}
\arguments{
\item{data}{The data parameter}

\item{ctrl}{The ctrl parameter}

\item{ctrlAll}{The ctrlAll parameter}

\item{statistic}{The statistic parameter}

\item{plot}{The plot parameter}

\item{delta}{The delta parameter}

\item{output}{The output parameter}

\item{ylim}{The ylim parameter}

\item{arrayType}{The arrayType parameter}
}
\value{
CNV calls for the HM450 routine
}
\description{
MethylMaster findSegments2 function
Function originally by Lucas Salas MD, PhD Dartmouth College 2021
}
