% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{GenStatM}
\alias{GenStatM}
\title{GenStatM.R}
\usage{
GenStatM(dnaM.m, pheno.v, chiptype = "450k")
}
\value{
A matrix of the infile
}
\description{
https://rdrr.io/bioc/FEM/src/R/GenStatM.R
}
