% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{getLRR.mm}
\alias{getLRR.mm}
\title{getLRR.mm}
\usage{
getLRR.mm(
  rgSet = rgset,
  Normals = Normals,
  sampNames = sampNames,
  QN = FALSE,
  Ref = "median",
  mode.bw = 0.1,
  mode.method = "naive",
  normal.cnv = TRUE,
  mean.center = TRUE,
  filter.autosomal.probes = FALSE,
  autosomal = NULL,
  filterProbes = FALSE,
  keep_fnobj = FALSE,
  ...
)
}
\arguments{
\item{rgSet}{The rgSet parameter}

\item{Normals}{The Normals parameter}

\item{sampNames}{The sampNames parameter}

\item{QN}{The QN parameter}

\item{Ref}{The Ref parameter}

\item{mode.bw}{The mode.bw parameter}

\item{mode.method}{The mode.method parameter}

\item{normal.cnv}{The normal.cnv parameter}

\item{mean.center}{The mean.center parameter}

\item{filter.autosomal.probes}{The filter.autosomal.probes parameter}

\item{autosomal}{The autosomal parameter}

\item{filterProbes}{The filterProbes parameter}

\item{...}{Additional parameter passed to getLRR.mm}

\item{import}{EpicopyData}
}
\value{
Return final LRR values
}
\description{
MethylMaster version of Epicopy getLRR function
}
