% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{infer_sex_karyotype}
\alias{infer_sex_karyotype}
\title{infer_sex_karyotype}
\usage{
infer_sex_karyotype(sset, ref.sset)
}
\arguments{
\item{sset}{SeSAMe sset}

\item{ref.sset}{SeSAMe reference sset}
}
\value{
sset with inferred sex data
}
\description{
Use sesame to infer sex karyotype
functionality by Jennifer Chen Dartmouth College 2021
}
