% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master.R
\name{methyl_master}
\alias{methyl_master}
\title{methyl_master}
\usage{
methyl_master(
  input.dir = NULL,
  output.dir = NULL,
  sample.sheet.path = NULL,
  file.sep = "/",
  r.lib.path = .libPaths()[1],
  n.cores = 1,
  os.type = "linux",
  proj = "TCGA-BLCA",
  visualize = FALSE,
  visualize.individual = FALSE,
  simplify.reduce = weightedmean,
  routine = "test",
  reference = NULL,
  reference.name = "all",
  split.by = NULL,
  comparison = NULL,
  overlap.density = 0.1,
  sesame.data.cache = "EPIC",
  sesame.data.normal = "EPIC.5.normal",
  sesame.ref.version = "hg38",
  hm450.workflow = "B",
  hm450.anno.file.path = NULL,
  champ.padj = 0.05,
  champ.control = FALSE,
  champ.run.combat = TRUE,
  champ.run.dmp = TRUE,
  champ.run.dmr = TRUE,
  champ.run.block = TRUE,
  champ.run.gsea = TRUE,
  champ.run.epimod = TRUE,
  epi.run.gistic = FALSE,
  compare.name = NULL,
  save.seg = FALSE,
  olaps.split.field = "Sample_ID",
  estimate.recurrence = FALSE,
  ov.less.stringent.ra.setting = ov.less.stringent.ra.setting,
  ov.pvalue = ov.pvalue,
  ov.keep.extra.columns = TRUE,
  ov.simplify.reduce = weightedmean,
  create.dir = FALSE,
  compare.list.files = FALSE,
  compare.files.in = NULL,
  ...
)
}
\value{
#NULL
}
\description{
Main function:
MultiMethylv1.0 : CNV calling from methylation data
Michael Mariani PhD Dartmouth College 2021
Possible routines:
"test"            ##Run a quick test
"process_sesame", ##Preprocess the TCGA and cord data in sesame format
"sesame",         ##Run Sesame  CNV calling (get segments)
"hm450" ,         ##Run 450K    CNV calling (get segments)
"champ" ,         ##Run ChAMP   CNV calling (get segments)
"epicopy" ,       ##Run EpiCopy CNV calling (get segments)
"compare"         ##Run algorithm comparison functionality
}
