% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_assay.R
\name{methyl_master_assay}
\alias{methyl_master_assay}
\title{##methyl_master_assay}
\usage{
methyl_master_assay(
  x,
  i = 1,
  withDimnames = TRUE,
  background = NA_integer_,
  sparse = FALSE
)
}
\value{
A reduced ReaggedExperiment assay object
}
\description{
MethylMasteR Assay function
Originally from sparseAssay() from the
RaggedExperiment Package
I have modified my own version
here for use in methyl master
}
