% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_compare.R
\name{methyl_master_compare}
\alias{methyl_master_compare}
\title{methyl_master_compare}
\usage{
methyl_master_compare(
  compare.list.files = FALSE,
  compare.files.in = NULL,
  compare.input.dir = NULL,
  compare.output.dir = NULL,
  ...
)
}
\value{
#outputs a comparison plot to <compare.output.dir>
}
\description{
compare results from different routine outputs
Michael Mariani PhD Dartmouth College 2021
}
