% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_custom.R
\name{methyl_master_custom}
\alias{methyl_master_custom}
\title{##methyl_master_custom}
\usage{
methyl_master_custom(
  custom.idat.files.dir = getwd(),
  custom.output.dir = getwd(),
  custom.ref = NULL,
  custom.sample.sheet.path = NULL,
  custom.comparison = NULL,
  custom.file.sep = NULL,
  custom.data.cache = "EPIC",
  custom.data.normal = "EPIC.5.normal",
  custom.ref.version = "hg38",
  custom.reference = "internal",
  custom.save.seg = FALSE,
  ...
)
}
\arguments{
\item{custom.idat.files.dir}{The input directory for the custom routine}

\item{custom.output.dir}{The ouptut directory for the custom routine}

\item{custom.sample.sheet.path}{The path to the MethylMaster sample sheet}

\item{custom.comparison}{the 2-element vector of Sample_Group levels to be
compared. First element is taken as the treatment and second as the control
if "reference" is set to "internal" the second element is ignored}

\item{custom.file.sep}{the file separator to use}

\item{custom.data.cache}{the custom data cache to use}

\item{custom.data.normal}{the custom normal data set to use,
e.g. Epic.5.Normal}

\item{custom.ref.version}{the custom reference version (default is hg38)}

\item{custom.reference}{the custom reference to use}

\item{custom.save.seg}{save the segmentation results as .RData object}

\item{...}{additional parameters to passs to methyl_master_custom}
}
\value{
A seg object for downstream analysis
}
\description{
MethylMasteR run custom function
}
