% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy.R
\name{methyl_master_epicopy}
\alias{methyl_master_epicopy}
\title{MethylMasteR Epicopy() version}
\usage{
methyl_master_epicopy(
  epi.target.dir = NULL,
  epi.output.dir = NULL,
  epi.single.file = TRUE,
  epi.single.file.path = NULL,
  epi.comparisons = NULL,
  epi.run.gistic = FALSE,
  epi.ncores = n.cores,
  epi.ref = "median",
  epi.less.stringent.ra = FALSE,
  epi.normals = "Sample_Group",
  epi.samp.names = NULL,
  epi.qn = FALSE,
  epi.mode.bw = 0.1,
  epi.mode.method = "naive",
  epi.normal.cnv = TRUE,
  epi.mean.center = TRUE,
  epi.filter.probes = FALSE,
  epi.retained.probes = NULL,
  epi.keepfnobj = FALSE,
  epi.fn.output = NULL,
  ...
)
}
\description{
MethylMasteR Epicopy() version

Lock updated epicopy functions in the environment:
load(paste0(work.dir,
file.sep,
"sesame_rgset_tumor.RData"))
load(paste0(work.dir,
file.sep,
"sesame_rgset_normal.RData"))
if(use.epi.normal==TRUE){
rgnormal <- sesame::SigSetsToRGChannelSet(sesame_ssets_normal)
}else{
load(paste0(work.dir,
file.sep,
"sesame_rgset_normal.RData"))
rgnormal <- rgset_normal
rm(rgset_normal)
}
Not run:
Run epicopy on included example TCGA data
input_loc <- system.file('extdata',
'raw_idat',
package = 'Epicopy')
No output directory, returns only local R object
}
