% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_hm450.R
\name{methyl_master_formatting_hm450}
\alias{methyl_master_formatting_hm450}
\title{methyl_master_formatting_hm450}
\usage{
methyl_master_formatting_hm450(
  hm450.form.seg = NULL,
  hm450.form.output.dir = getwd(),
  hm450.form.sample.sheet.path = NULL,
  hm450.form.reference = "internal",
  hm450.form.split.by = NULL,
  hm450.form.workflow = "B",
  hm450.form.comparison = NULL,
  hm450.form.save.seg = FALSE,
  hm450.form.anno.file.path,
  ...
)
}
\arguments{
\item{hm450.form.seg}{The hm450 routine seg results as input to be formatted}

\item{hm450.form.output.dir}{The output dir for the formatting results}

\item{hm450.form.sample.sheet.path}{The MehtylMaster sample sheet path}

\item{hm450.form.reference}{The hm450 reference that was used}

\item{hm450.form.split.by}{The split.by sample sheet field that was used}

\item{hm450.form.workflow}{The specific HM450 workflow that was used}

\item{hm450.form.comparison}{The MethylMaster two-element comparison vector}

\item{hm450.form.save.seg}{Whether to save the formatted HM450 results}

\item{hm450.form.anno.file.path}{The hm450 annotation file path}

\item{...}{Additional parameters passed to methyl_master_formatting_hm450}
}
\value{
#Formatted seg list object for overlaps and visualizing
}
\description{
Formatting the hm450 output to prepare for plotting etc.
}
