% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_sesame.R
\name{methyl_master_formatting_sesame}
\alias{methyl_master_formatting_sesame}
\title{methyl_master_formatting_sesame}
\usage{
methyl_master_formatting_sesame(
  sesame.form.seg,
  sesame.form.output.dir,
  sesame.form.sample.sheet.path,
  sesame.form.reference,
  sesame.form.split.by,
  sesame.form.comparison,
  sesame.form.save.seg,
  ...
)
}
\value{
#Formatted seg list object for visualizing
}
\description{
Main function:
MultiMethylv1.0 : CNV calling from methylation data
Michael Mariani PhD Dartmouth College 2021
Possible routines:
"test"            ##Run a quick test
"process_sesame", ##Preprocess the TCGA and cord data in sesame format
"sesame",         ##Run Sesame  CNV calling (get segments)
"hm450" ,         ##Run 450K    CNV calling (get segments)
"champ" ,         ##Run ChAMP   CNV calling (get segments)
"epicopy" ,       ##Run EpiCopy CNV calling (get segments)
"compare"         ##Run algorithm comparison functionality
}
