% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_k450.R
\name{methyl_master_hm450}
\alias{methyl_master_hm450}
\title{My 450k}
\usage{
methyl_master_hm450(
  hm450.idat.files.dir = NULL,
  hm450.sample.sheet.path = NULL,
  hm450.comparison = c("tumor", "normal"),
  hm450.split.by = NULL,
  hm450.reference = "internal",
  hm450.sesame.data.cache = "EPIC",
  hm450.sesame.data.normal = "EPIC.5.normal",
  hm450.sesame.ref.version = "hg38",
  ...
)
}
\value{
Nothing
}
\description{
My version of 450k analyses
}
