% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_load_data_objects.R
\name{methyl_master_load_data_objects}
\alias{methyl_master_load_data_objects}
\title{methyl_master_load_data_objects}
\usage{
methyl_master_load_data_objects(files.dir, file.sep)
}
\arguments{
\item{files.dir}{The parameter containing the R files}

\item{file.sepThe}{the desired file separator}
}
\value{
loaded data object(s)
}
\description{
Possible function for loading data objects
required for MethylMaster
Load required data objects:
load(file=paste0(files.dir,"/","hm450.manifest.hg38.rda"))
annotation_df <- as.data.frame(
data.table::fread(
paste0(work.dir,
file.sep,
"HM450.hg38.manifest.tsv")
)
)
}
