% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_plot_individual.R
\name{methyl_master_plot_individual}
\alias{methyl_master_plot_individual}
\title{methyl_master_plot_individual}
\usage{
methyl_master_plot_individual(
  pi.seg = NULL,
  pi.output.dir = NULL,
  pi.name = NULL,
  ...
)
}
\value{
#NULL
}
\description{
output signal inensity plots for individual samples
##Michael Mariani PhD Dartmouth College 2021
}
