% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{methyl_master_population_ranges}
\alias{methyl_master_population_ranges}
\title{methyl_master_population_ranges}
\usage{
methyl_master_population_ranges(
  grl,
  mode = c("density", "RO"),
  density = 0.1,
  ro.thresh = 0.5,
  multi.assign = FALSE,
  verbose = FALSE,
  min.size = 2,
  classify.ranges = TRUE,
  type.thresh = 0.1,
  est.recur = FALSE
)
}
\arguments{
\item{grl}{The grl parameter}

\item{mode}{The mode parameter}

\item{density}{The density parameter}

\item{ro.thresh}{The ro.thresh parameter}

\item{multi.assign}{The multi.assign parameter}

\item{verbose}{The verbose parameter}

\item{min.size}{The min.size parameter}

\item{classify.ranges}{The classify.ranges parameter}

\item{type.thresh}{The type.thresh parameter}

\item{est.recur}{The est.recur parameter}

\item{data}{The data parameter}
}
\value{
A populationRanges CNV overlaps object
}
\description{
MwthylMaster version of the populatonRanges function
originally by Ludwig Geistlinger 2017
}
