% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_prepare_sample_sheet.R
\name{methyl_master_prepare_sample_sheet}
\alias{methyl_master_prepare_sample_sheet}
\title{methyl_master_prepare_sample_sheet}
\usage{
methyl_master_prepare_sample_sheet(
  prep.output.path.name = NULL,
  prep.file.sep = NULL,
  prep.idat.dir = NULL,
  prep.sample.sheet = NULL,
  prep.primary.samples = NULL,
  prep.sample.group = NULL,
  prep.platform = NULL,
  prep.batch = 1
)
}
\arguments{
\item{prep.output.path.name}{The output file destination}

\item{prep.file.sep}{The file separator to use}

\item{prep.idat.dir}{The .idat file directory}

\item{prep.sample.sheet}{The sample sheet to use}

\item{prep.primary.samples}{The specified primary sample IDs}

\item{prep.sample.group}{The specified sample groups}

\item{prep.platform}{The platform}

\item{prep.batch}{the batch variable}
}
\value{
A sample sheet for use with MethylMaster
}
\description{
compare results from different routine outputs
For example, create a MethylMaster style sample sheet
from selected 3p-del (positive control) KIRC samples
from the firehose collection on CBioPortal
}
