% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_qreduceassay.R
\name{methyl_master_qreduceassay}
\alias{methyl_master_qreduceassay}
\title{##methyl_master_qreduceassay}
\usage{
methyl_master_qreduceassay(
  x,
  query,
  simplifyReduce,
  i = 1,
  withDimnames = TRUE,
  background = NA_integer_
)
}
\value{
A reduced ReaggedExperiment assay object
}
\description{
MethylMasteR qreduceAssay function
Originally from the RaggedExperiment Package
I have modified my own version here for debugging
and tighter control
}
