% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_select_random_samples.R
\name{methyl_master_select_random_samples}
\alias{methyl_master_select_random_samples}
\title{methyl_master_load_data_objects}
\usage{
methyl_master_select_random_samples(normal.path, tumor.path)
}
\arguments{
\item{normal.path}{The normal path}

\item{tumor.path}{The tumor path}
}
\value{
Return randomly sampled TCGA IDAT samples
}
\description{
Possible function for loading data objects
Michael Mariani PhD Dartmouth College 2021
norm.path<-"G:\\My Drive\\dartmouth\\salas_lab_working\\cnv\\BLCA_normal.txt"
tum.path<-"G:\\My Drive\\dartmouth\\salas_lab_working\\cnv\\BLCA_normal.txt"
}
