% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_tcga_clin_data.R
\name{methyl_master_tcga_clin_data}
\alias{methyl_master_tcga_clin_data}
\title{methyl_master_tcga_clin_data}
\usage{
methyl_master_tcga_clin_data(
  clin.tcga.file.path = NULL,
  clin.sample.names = NULL,
  clin.copy.files = FALSE,
  clin.sample.sheet = TRUE,
  clin.idat.dir = getwd(),
  clin.sub.dir = NULL
)
}
\arguments{
\item{clin.tcga.file.path}{The path to the tcga files}

\item{clin.sample.names}{The sample names}

\item{clin.copy.files}{Whether to copy the files to a different directory}

\item{clin.sample.sheet}{Whether to use a sample sheet specifying files}

\item{clin.idat.dir}{The idat dir}

\item{clin.sub.dir}{The sub directory}
}
\value{
#outputs a vector or excel of
primary ids for input tcga smamples names
}
\description{
get primary ids from sample names from tcga clinical file
}
