% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_get_cord_sex.R
\name{methyl_measter_get_cord_sex}
\alias{methyl_measter_get_cord_sex}
\title{methyl_master_get_cord_sex}
\usage{
methyl_measter_get_cord_sex(
  mmcs.idat.files.dir = getwd(),
  mmcs.platform = "EPIC",
  mmcs.output.file.name = NULL
)
}
\arguments{
\item{mmcs.idat.files.dir}{The .idat files directory}

\item{mmcs.platform}{The platform}

\item{mmcs.output.file.name}{The output file name}
}
\value{
Outputs a file of the sexes for select cord samples
}
\description{
Get sex information for cord samples
}
