% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_population_ranges.R
\name{plotRecurrentRegions}
\alias{plotRecurrentRegions}
\title{plotRecurrentRegions}
\usage{
plotRecurrentRegions(regs, genome, chr, pthresh = 0.05)
}
\arguments{
\item{regs}{The regs Parameter}

\item{genome}{The genome parameter}

\item{chr}{The chr parameter}

\item{pthresh}{The pthresh parameter}
}
\description{
MethylMasteR version of plotReccurentRegions function
"Plot recurrent CNV regions
Illustrates summarized CNV regions along a chromosome."
}
