% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{read.metharray.mm}
\alias{read.metharray.mm}
\title{readmetharray.mm}
\usage{
read.metharray.mm(basenames, extended = FALSE, verbose = FALSE, force = FALSE)
}
\arguments{
\item{basenames}{The basenames parameter}

\item{extended}{The extended parameter}

\item{verbose}{The verbose parameter}

\item{force}{The force parameter}

\item{diff.array.type}{The diff.array.type parameter}
}
\value{
Methylation array data
}
\description{
MethylMaster version of Minfi read.metharray() function
}
