% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{read.metharray.sheet.mm}
\alias{read.metharray.sheet.mm}
\title{read.metharray.sheet.mm}
\usage{
read.metharray.sheet.mm(
  base,
  pattern = "csv$",
  ignore.case = TRUE,
  recursive = TRUE,
  single.file = TRUE,
  single.file.path = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{base}{The base parameter}

\item{pattern}{The pattern parameter}

\item{ignore.case}{The ignore.case parameter}

\item{recursive}{The recursive parameter}

\item{single.file}{The single.file parameter}

\item{single.file.path}{The single.file.path parameter}

\item{verbose}{The verbose parameter}

\item{...}{Additional arguments passed to read.metharray.mm}
}
\value{
A formatted csv as a dataframe
}
\description{
MethylMaster version of read.metharray.sheet() from minfi
}
