% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_CopyNumber450kCancer.R
\name{ReadData}
\alias{ReadData}
\title{CopyNumber450kCancer.R}
\usage{
ReadData(regions_file, Sample_list, copynumber450k = FALSE)
}
\value{

}
\description{
#################### CopyNumber450kCancer ###################################
############### Note the offset function ####################################
#############################################################################
#############################################################################
CopyNumber450kCancer.R
Function to read the data------
Functions for checking the headers:::
this function to check the header of the input file:
("Sample" "Chromosome" "bp.Start"  "bp.End" "Num.of.Markers" "Mean")
}
