% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{readSheet.mm}
\alias{readSheet.mm}
\title{readSheet.mm}
\usage{
readSheet.mm(file, recursive = TRUE)
}
\arguments{
\item{file}{The file parameter}

\item{recursive}{The recursive parameter}
}
\value{
Formatted csv as dataframe
}
\description{
MethylMaster version of readSheet() from Minfi
}
