% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_plot_individual.R
\name{visualizeSegments.mm}
\alias{visualizeSegments.mm}
\title{visualizeSegments.mm}
\usage{
visualizeSegments.mm(seg, to.plot = NULL)
}
\arguments{
\item{seg}{The sesame segmentation object}

\item{to.plot}{the to.plot parameter}
}
\value{
A ggplot
}
\description{
Methyl master modified sesame::visualizeSegments function
from SeSAMe
}
