% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{binding_frames_mm}
\alias{binding_frames_mm}
\title{Bind individual sesame frames together}
\usage{
binding_frames_mm(x)
}
\arguments{
\item{x}{sesame-style data frame}
}
\value{
segment dataframe output of multiple dfs bound together
}
\description{
Bind individual sesame frames together
}
