% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{calc_seg_state}
\alias{calc_seg_state}
\title{calc_seg_state}
\usage{
calc_seg_state(seg.means = NULL, upper.thresh = 4, cutoff = NULL)
}
\arguments{
\item{seg.means}{vector of seg.means}

\item{upper.thresh}{CN state threshold, any number above this will be
assigned this value}

\item{cutoff}{vector of cutoff threshold for loss and
gain respectively}
}
\value{
vector of calculated CNV states
}
\description{
common function used to calculate seg state from seg mean
}
