% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.load.mm}
\alias{champ.load.mm}
\title{champ.load.mm}
\usage{
champ.load.mm(
  directory = getwd(),
  sample.sheet = NULL,
  compare.group = compare.group,
  control = FALSE,
  method = "ChAMP",
  methValue = "B",
  autoimpute = TRUE,
  filterDetP = TRUE,
  ProbeCutoff = 0,
  SampleCutoff = 0.1,
  detPcut = 0.01,
  filterBeads = TRUE,
  beadCutoff = 0.05,
  filterNoCG = TRUE,
  filterSNPs = TRUE,
  population = NULL,
  filterMultiHit = TRUE,
  filterXY = TRUE,
  force = FALSE,
  arraytype = "450K"
)
}
\arguments{
\item{directory}{The input directory}

\item{sample.sheet}{The MethylMaster sample sheet}

\item{compare.group}{The MethylMaster comparison vector}

\item{control}{Whether to specify control or not}

\item{method}{Specify the method parameter}

\item{methValue}{Specify the methValue parameter}

\item{autoimpute}{Specify the autoimpute parameter}

\item{filterDetP}{Specify the filterDetP parameter}

\item{ProbeCutoff}{Specify the ProbeCutoff parameter}

\item{SampleCutoff}{Specify the SampleCutoff parameter}

\item{detPcut}{Specify the detPcut parameter}

\item{filterBeads}{Specify the filterBeads parameter}

\item{beadCutoff}{Specify the beadCutoff parameter}

\item{filterNoCG}{Specify the NoCG parameter}

\item{filterSNPs}{Specify the filterSNPs parameter}

\item{population}{Specify the population parameter}

\item{filterMultiHit}{Specify the filterMultiHit parameter}

\item{filterXY}{Specify the filterXY parameter}

\item{force}{Specify the force parameter}

\item{arraytype}{Specify the arraytype parameter}
}
\value{
Return a list of loaded data
}
\description{
MethylMaster modified champ.load() function
}
