% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.process.mm}
\alias{champ.process.mm}
\title{champ.process.mm}
\usage{
champ.process.mm(
  runload = TRUE,
  directory = getwd(),
  resultsDir = "./CHAMP_RESULT/",
  sample.sheet = NULL,
  arraytype = "450K",
  controlGroup = "champCtls",
  filters = c("XY", "DetP", "Beads", "NoCG", "SNP", "MultiHit"),
  runimpute = TRUE,
  imputemethod = "Combine",
  runQC = TRUE,
  QCplots = c("mdsPlot", "densityPlot", "dendrogram"),
  runnorm = TRUE,
  normalizationmethod = "BMIQ",
  runSVD = TRUE,
  RGEffect = FALSE,
  runCombat = TRUE,
  batchname = c("Slide"),
  runDMP = TRUE,
  runDMR = TRUE,
  DMRmethod = "Bumphunter",
  runBlock = TRUE,
  runGSEA = TRUE,
  runEpiMod = TRUE,
  runCNA = TRUE,
  control = TRUE,
  runRefBase = FALSE,
  compare.group = NULL,
  adjPVal = 0.05,
  PDFplot = TRUE,
  Rplot = TRUE,
  cores = 1,
  saveStepresults = TRUE
)
}
\value{
Return a champ.result object
}
\description{
champ.process function modified by
Michael Mariani PhD Dartmouth College 2021
}
