% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{champ.QC}
\alias{champ.QC}
\title{champ.QC()}
\usage{
champ.QC(
  beta = NULL,
  pheno = NULL,
  mdsPlot = NULL,
  densityPlot = NULL,
  dendrogram = NULL,
  PDFplot = NULL,
  Rplot = NULL,
  Feature.sel = NULL,
  resultsDir = NULL
)
}
\arguments{
\item{infile}{Path to the input file}

\item{champ.beta}{champ.beta parameter}

\item{champ.pheno}{champ.pheno parameter}

\item{champ.mds.plot}{champ.mds.plot parameter}

\item{champ.density.plot}{champ.density.plot parameter}

\item{champ.dendrogram}{champ.dendrogram parameter}

\item{champ.pdf.plot}{champ.pdf parameter}

\item{champ.rplot}{champ.rplot parameter}

\item{champ.feature.sel}{champ.feature.sel parameter}

\item{champ.results.dir}{champ.results.dir parameter}
}
\value{
A matrix of the infile
}
\description{
MethylMaster version of the champ.QC function()
}
