% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{DoEpiMod}
\alias{DoEpiMod}
\title{DoEpiMod}
\usage{
DoEpiMod(
  intEpi.o,
  nseeds = 100,
  gamma = 0.5,
  nMC = 1000,
  sizeR.v = c(1, 100),
  minsizeOUT = 10,
  writeOUT = TRUE,
  nameSTUDY = "X",
  ew.v = NULL
)
}
\value{
A matrix of the infile
}
\description{
https://rdrr.io/bioc/FEM/src/R/DoEpiMod.R
}
