% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{DoIntEpi450k}
\alias{DoIntEpi450k}
\title{DoIntEpi450k}
\usage{
DoIntEpi450k(statM.o, adj.m, c, dmaMode = "avbeta")
}
\value{
A matrix of the infile
}
\description{
##https://rdrr.io/bioc/FEM/src/R/DoIntEpi450k.R
}
