% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{.funnorm.mm}
\alias{.funnorm.mm}
\title{Epciopy modified .funnorm function}
\usage{
.funnorm.mm(rgSet, nPCs = 2, sex = NULL, verbose = TRUE)
}
\value{
return final gm set
}
\description{
Epciopy modified .funnorm function
}
