% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{fast_450k_binning}
\alias{fast_450k_binning}
\title{fast_450_k_binning}
\usage{
fast_450k_binning(
  proc.samples,
  med.ref.values,
  proc.ref.samples,
  bin.size = 5e+05
)
}
\arguments{
\item{proc.samples}{The samples used as input}

\item{med.ref.values}{The median reference values}

\item{proc.ref.samples}{the refernce samples}

\item{bin.size}{The bin size}
}
\value{
A 450k-style candidates data frame
}
\description{
A faster version for binding 450K bins
Originally function by Knoll et al., 2017
}
