% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{FemModShow}
\alias{FemModShow}
\title{FemModShow}
\usage{
FemModShow(mod, name, fem.o, mode = "Integration")
}
\value{
A matrix of the infile
}
\description{
https://rdrr.io/bioc/FEM/src/R/FemModShow.R
}
