% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_findSegments2.R
\name{findSegments2}
\alias{findSegments2}
\title{Modified findSegments function}
\usage{
findSegments2(
  data,
  ctrl,
  ctrlAll,
  statistic = "wilcoxon",
  plot = FALSE,
  delta = 500,
  output = "diff",
  ylim = NULL,
  arrayType = "auto"
)
}
\value{

}
\description{
Find segements
}
