% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_champ_helper_functions.R
\name{GenStatM}
\alias{GenStatM}
\title{GenStatM.R}
\usage{
GenStatM(dnaM.m, pheno.v, chiptype = "450k")
}
\arguments{
\item{dnaM.m}{The dnaM.m parameter}

\item{pheno.v}{The pheno.v parameter}

\item{chiptype}{The chiptype parameter}
}
\value{
A results list
}
\description{
MethylMaster of the GenStatM.R function see:
https://rdrr.io/bioc/FEM/src/R/GenStatM.R
}
