% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{getLRR.mm}
\alias{getLRR.mm}
\title{Epciopy modified getLRR function}
\usage{
getLRR.mm(
  rgSet = rgset,
  Normals = Normals,
  sampNames = sampNames,
  QN = FALSE,
  Ref = "median",
  mode.bw = 0.1,
  mode.method = "naive",
  normal.cnv = TRUE,
  mean.center = TRUE,
  filter.autosomal.probes = FALSE,
  autosomal = NULL,
  filterProbes = FALSE,
  keep_fnobj = FALSE,
  ...
)
}
\value{
return final LRR values
}
\description{
Epciopy modified getLRR function
}
