% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{harmonized_rg_sets}
\alias{harmonized_rg_sets}
\title{Harmonize RG data}
\usage{
harmonized_rg_sets(rgset1, rgset2)
}
\arguments{
\item{rgset1}{Path to the input file}
}
\value{
Harmonized data set
}
\description{
Harmonize RG data
}
