% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{infer_sex_karyotype}
\alias{infer_sex_karyotype}
\title{Use sesame to infer sex karyotype}
\usage{
infer_sex_karyotype(sset, ref.sset)
}
\arguments{
\item{sset}{Path to the input file}
}
\value{
sset with inferred sex data
}
\description{
Use sesame to infer sex karyotype
}
