% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{LRRtoCNA.mm}
\alias{LRRtoCNA.mm}
\title{LRRtoCNA.mm}
\usage{
LRRtoCNA.mm(
  LRR,
  sampID = NULL,
  ncores = 1L,
  seg.undo.splits = "sdundo",
  seg.undo.SD = 2,
  epic.manifest.path = getwd(),
  platform = c("hm450", "EPIC"),
  ...
)
}
\arguments{
\item{LRR}{The LRR parameter}

\item{sampID}{The sampID parameter}

\item{ncores}{The ncores parameter, note that multiple cores may not work
on every system}

\item{seg.undo.splits}{The seg.undo.splits parameter}

\item{seg.undo.SD}{The seg.undo.SD parameter}

\item{epic.manifest.path}{The epic.manifest.path parameter}

\item{platform}{The platform parameter}
}
\value{
Returns CNA segments
}
\description{
MethylMaster version of the Epicopy LRRtoCNA function
Epicopy convert log r ratio data to cna calls:
Similar to Lucas Salas' PhD modified LRRtoCNA2 function
}
