% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{LRRtoCNA.mm}
\alias{LRRtoCNA.mm}
\title{Epciopy modified convert log r ratio data to cna calls:}
\usage{
LRRtoCNA.mm(
  LRR,
  sampID = NULL,
  ncores = 1L,
  seg.undo.splits = "sdundo",
  seg.undo.SD = 2,
  platform = c("hm450", "EPIC"),
  ...
)
}
\value{
returns segments
}
\description{
Same as LucasLRRtoCNA2 function
}
