% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMasteR.R
\name{methyl_master}
\alias{methyl_master}
\title{Main functions:}
\usage{
methyl_master(
  input.dir = input.dir,
  output.dir = output.dir,
  sample.sheet.path = sample.sheet.path,
  r.lib.path = .libPaths()[1],
  n.cores = 1,
  os.type = "linux",
  proj = "TCGA-BLCA",
  visualize = FALSE,
  visualize.individual = FALSE,
  weighted.mean = "normal",
  routine = "test",
  reference = NULL,
  split.by = NULL,
  comparisons = NULL,
  overlap.density = 0.1,
  epi.run.gistic = FALSE,
  ...
)
}
\value{
#NULL
}
\description{
MultiMethylv1.0 : CNV calling from methylation data
Michael Mariani PhD Dartmouth College 2021
Possible routines:
"test"            ##Run a quick test
"process_sesame", ##Preprocess the TCGA and cord data in sesame format
"sesame",         ##Run Sesame  CNV calling (get segments)
"hm450" ,         ##Run 450K    CNV calling (get segments)
"champ" ,         ##Run ChAMP   CNV calling (get segments)
"epicopy" ,       ##Run EpiCopy CNV calling (get segments)
"compare"         ##Run algorithm comparison functionality
}
