% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master.R
\name{methyl_master}
\alias{methyl_master}
\title{methyl_master}
\usage{
methyl_master(
  input.dir = NULL,
  output.dir = NULL,
  sample.sheet.path = NULL,
  file.sep = "/",
  r.lib.path = .libPaths()[1],
  n.cores = 1,
  os.type = "linux",
  proj = "TCGA-BLCA",
  visualize = FALSE,
  visualize.individual = FALSE,
  simplify.reduce = weightedmean,
  routine = "test",
  reference = NULL,
  reference.name = "all",
  split.by = NULL,
  comparison = NULL,
  form.thresholds = NULL,
  overlap.density = 0.1,
  sesame.data.cache = "EPIC",
  sesame.data.normal = "EPIC.5.normal",
  genome.version = "hg38",
  hm450.workflow = "B",
  champ.padj = 0.05,
  champ.control = FALSE,
  champ.run.combat = TRUE,
  champ.run.dmp = TRUE,
  champ.run.dmr = TRUE,
  champ.run.block = TRUE,
  champ.run.gsea = TRUE,
  champ.run.epimod = TRUE,
  epi.run.gistic = FALSE,
  compare.name = NULL,
  save.seg = FALSE,
  olaps.split.field = "Sample_ID",
  estimate.recurrence = FALSE,
  ov.less.stringent.ra.setting = ov.less.stringent.ra.setting,
  ov.pvalue = ov.pvalue,
  ov.keep.extra.columns = TRUE,
  ov.simplify.reduce = weightedmean,
  create.dir = FALSE,
  compare.list.files = FALSE,
  compare.files.in = NULL,
  compare.names = NULL,
  compare.olaps.size = 1,
  ...
)
}
\arguments{
\item{input.dir}{Input directory}

\item{output.dir}{Output directory}

\item{sample.sheet.path}{Path to sample sheet}

\item{file.sep}{File separator}

\item{r.lib.path}{Path to r library default is .libPaths()}

\item{n.cores}{Number of cores to run}

\item{os.type}{The os type running the software}

\item{proj}{The project name}

\item{visualize}{Visualize the output}

\item{visualize.individual}{Visualize individual sample plots for sesame}

\item{simplify.reduce}{The reduction function being applied to the output}

\item{routine}{The specific routine to run}

\item{reference}{The reference to use}

\item{reference.name}{The reference name, for the epicopy routine}

\item{split.by}{Which additional metadata column to split the analysis by}

\item{comparison}{Which groups from the metadata "Sample_Group" to compare}

\item{form.thresholds}{The thresholds for cnv state calling}

\item{overlap.density}{The overlap density applied to final results from
the modified population ranges function}

\item{sesame.data.cache}{The sesame data cache to use}

\item{sesame.data.normal}{Which sesame normal data to use e.g."EPIC.5.Normal"}

\item{genome.version}{The genome version}

\item{hm450.workflow}{which hm450 workflow to use with the hm450 routine:
"A", "B", or "C"}

\item{champ.padj}{adjusted p-value threshold for the ChAMP routine}

\item{champ.control}{select the champ control to use}

\item{champ.run.combat}{run combat batch correction in the champ routine}

\item{champ.run.dmp}{run dmp in the champ routin}

\item{champ.run.dmr}{run dmr in the champ routine}

\item{champ.run.block}{run block in the champ routine}

\item{champ.run.gsea}{run gsea in the champ routine}

\item{champ.run.epimod}{run epimod in the champ routine}

\item{epi.run.gistic}{run GISTIC in the epicopy routine}

\item{compare.name}{the compare.name to use}

\item{save.seg}{whether to save the seg output or not}

\item{olaps.split.field}{which field in the olaps to split the results by
the default (recommended) is to use "Sample_ID"}

\item{estimate.recurrence}{whether to use the estimate.recurrence feature
in populationsRanges when looking at overlaps within a particular routine}

\item{ov.less.stringent.ra.setting}{whether to use less stringent
method when comparing overlaps of cnvs from an individual routine}

\item{ov.pvalue}{the pvalue cutoff to use within a routine}

\item{ov.keep.extra.columns}{whether to keep extra meta data columns or
not when collating results from an individual routine}

\item{ov.simplify.reduce}{which reduction method to use to compare overlaps
within a routine}

\item{create.dir}{create output directory when running MethylMaster,
if already exists it will be over-written}

\item{compare.list.files}{whether to list the files individually to be
compared with the compare routine or not, otherwise will search through
the input.dir path}

\item{compare.files.in}{vector of file paths to individual routine results
to use with the compare routine}

\item{compare.names}{the names of the files (with paths) to run with the
compare routine}

\item{compare.olaps.size}{the overlaps size to use when considering an
overlap hit across multiple routine results (using the compare routin)}

\item{...}{additional parameters to be passed to methyl_master}
}
\description{
MethyMaster main function:
CNV calling platform from methylation data algorithms
and additional comparison functionality
Michael Mariani PhD Dartmouth College 2021-2022
Possible routines:
"test"            ##Run a quick test
"process_sesame", ##Preprocess the TCGA and cord data in sesame format
"sesame",         ##Run Sesame  CNV calling (get segments)
"hm450" ,         ##Run 450K    CNV calling (get segments)
"champ" ,         ##Run ChAMP   CNV calling (get segments)
"epicopy" ,       ##Run EpiCopy CNV calling (get segments)
"compare"         ##Run algorithm comparison functionality
}
