% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methy_master_champ.R
\name{methyl_master_champ}
\alias{methyl_master_champ}
\title{Run ChAMP}
\usage{
methyl_master_champ(
  champ.directory,
  champ.array.type = "450K",
  champ.batch.name = c("batch"),
  champ.padj = 0.05,
  champ.ncores = n.cores,
  champ.control = TRUE,
  champ.control.group = "normal",
  champ.runimpute = TRUE,
  champ.runQC = TRUE,
  champ.runnorm = TRUE,
  champ.runSVD = TRUE,
  champ.runCombat = TRUE,
  champ.runDMP = TRUE,
  champ.runDMR = TRUE,
  champ.runBlock = TRUE,
  champ.runGSEA = TRUE,
  ...
)
}
\value{
#NULL
}
\description{
The full pipeline can be run at once with one command:
Testing:
data(EPICSimData)
champ.test.directory <- system.file("extdata",
package="ChAMPdata")
myLoad.test <- champ.load(champ.test.directory,
arraytype="450K")
myLoad.test$pd$Sample_Name
myLoad.test$pd$Sample_Group
champ.results <- champ.process(directory=champ.test.directory)
myLoad <- champ.load(idat.pooled.files.dir,
arraytype="450K")
myLoad$pd$Pool_ID
myLoad$pd$Pool_ID
myLoad$pd$Pool_ID
myLoad$pd$Sample_Group
}
\details{
Can test ChAMP with the test data: data(EPICSimData)
}
