% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_compare.R
\name{methyl_master_compare}
\alias{methyl_master_compare}
\title{methyl_master_compare}
\usage{
methyl_master_compare(
  compare.list.files = FALSE,
  compare.files.in = NULL,
  compare.input.dir = getwd(),
  compare.output.dir = getwd(),
  compare.names = NULL,
  compare.olaps.size = 1,
  ...
)
}
\arguments{
\item{compare.list.files}{list files of not if set to FALSE, will recursively
search the compare.input.dir for overlaps_filt and time_mem files to compare}

\item{compare.files.in}{vector of individual files in for the compare routine}

\item{compare.input.dir}{input directory for the compare routine}

\item{compare.output.dir}{output directory for the compare routine}

\item{compare.names}{the names of the results for the individual routines
that are being compared. Default is the routine results dir basename}

\item{compare.olaps.size}{size of overlaps across routines considered a hit
default is 1 (1 or more bp overlap)}

\item{...}{additional parameters to pass to methyl_master_compare}
}
\value{
Outputs a comparison plot to <compare.output.dir>
}
\description{
compare results from different routine outputs
Michael Mariani PhD Dartmouth College 2021
}
