% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy.R
\name{methyl_master_epicopy}
\alias{methyl_master_epicopy}
\title{methyl_master_epicopy}
\usage{
methyl_master_epicopy(
  epi.input.dir,
  epi.output.dir,
  epi.single.file,
  epi.single.file.path,
  epic.manifest.path,
  epi.reference.group,
  epi.comparison,
  epi.split.by,
  epi.ncores,
  epi.ref,
  epi.normals,
  epi.samp.names,
  epi.qn,
  epi.mode.bw,
  epi.mode.method,
  epi.normal.cnv,
  epi.mean.center,
  epi.filter.probes,
  epi.retained.probes,
  epi.keepfnobj,
  epi.fn.output,
  epi.run.gistic = FALSE,
  ...
)
}
\value{
epicopy results data frame
}
\description{
my modfiied epicopy function for MethylMaster
}
