% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_champ.R
\name{methyl_master_formatting_champ}
\alias{methyl_master_formatting_champ}
\title{methyl_master_formatting_champ}
\usage{
methyl_master_formatting_champ(
  champ.form.seg = NULL,
  champ.form.output.dir = getwd(),
  champ.form.save.seg = FALSE,
  champ.form.comparison = NULL,
  champ.form.padj = NULL
)
}
\value{
Formatted champ result as MethylMaster seg object
for downstream analysis
}
\description{
formatting the champ seg results for
overlaps and visualization etc.a
Michael Mariani PhD Dartmouth College 2021
}
