% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_epicopy.R
\name{methyl_master_formatting_epicopy}
\alias{methyl_master_formatting_epicopy}
\title{methyl_master_formatting_epicopy}
\usage{
methyl_master_formatting_epicopy(
  epi.form.seg = NULL,
  epi.form.output.dir = getwd(),
  epi.form.save.seg = FALSE,
  epi.form.comparison = NULL
)
}
\value{
Formatted epicopy result as MethylMaster seg object
for downstream analysis
}
\description{
Formatting Epicopy output into MethylMaster seg object
for downstream analyses
}
