% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_hm450.R
\name{methyl_master_formatting_hm450}
\alias{methyl_master_formatting_hm450}
\title{methyl_master_formatting_hm450}
\usage{
methyl_master_formatting_hm450(
  hm450.form.seg = NULL,
  hm450.form.output.dir = getwd(),
  hm450.form.sample.sheet.path = NULL,
  hm450.form.reference = "internal",
  hm450.form.split.by = NULL,
  hm450.form.workflow = "B",
  hm450.form.comparison = NULL,
  hm450.form.save.seg = FALSE,
  hm450.form.anno.file.path,
  ...
)
}
\value{
#Formatted seg list object for visualizing
}
\description{
formatting the hm450 output to prepare for plotting etc.
Michael Mariani PhD Dartmouth College 2021
}
