% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_formatting_sesame.R
\name{methyl_master_formatting_sesame}
\alias{methyl_master_formatting_sesame}
\title{methyl_master_formatting_sesame}
\usage{
methyl_master_formatting_sesame(
  sesame.form.seg = NULL,
  sesame.form.output.dir = getwd(),
  sesame.form.sample.sheet.path = NULL,
  sesame.form.reference = NULL,
  sesame.form.split.by = NULL,
  sesame.form.comparison = NULL,
  sesame.form.save.seg = FALSE,
  sesame.form.plot.individual = FALSE,
  sesame.form.thresholds = NULL,
  ...
)
}
\arguments{
\item{sesame.form.seg}{The input sesame segmentation list for formatting}

\item{sesame.form.output.dir}{The output dir for sesame formatting}

\item{sesame.form.reference}{The sesame reference used in the analysis}

\item{sesame.form.split.by}{The split.by field used in the sesame routine
in the MethylMaster sample sheet}

\item{sesame.form.comparison}{The MethylMaster comparison vector used in the
Sesame routine analysis}

\item{sesame.form.save.seg}{Whether or not to save the formatted seg output}

\item{sesame.form.plot.individual}{Whether to plot the individual sesame
sample signal plots}

\item{sesame.form.thresholds}{The thresholds used to determine the CNV
state, if NULL, the equation  seg.state <- round(2^seg.means * 2) is used}

\item{...}{additional parameters to pass to methyl_master_formatting_sesame}

\item{sample.form.sample.sheet.path}{The path to the MethylMaster
sample sheet}
}
\value{
Formatted seg list object for visualizing etc.
}
\description{
formatting the results of the sesame functionality to prepare
for comparison and output
}
