% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_hm450.R
\name{methyl_master_hm450}
\alias{methyl_master_hm450}
\title{methyl_master_hm450}
\usage{
methyl_master_hm450(
  hm450.input.dir = NULL,
  hm450.output.dir = NULL,
  hm450.sample.sheet.path = NULL,
  hm450.comparison = c("tumor", "normal"),
  hm450.split.by = NULL,
  hm450.reference = "internal",
  hm450.workflow = "B",
  hm450.file.sep = "/",
  hm450.sesame.data.cache = "EPIC",
  hm450.sesame.data.normal = "EPIC.5.normal",
  hm450.sesame.ref.version = "hg38",
  hm450.save.seg = FALSE,
  ...
)
}
\value{
#seg.out
}
\description{
My version of hm450 analyses
############ Samples and ref also need to be in same order ##########
}
