% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_hm450.R
\name{methyl_master_hm450}
\alias{methyl_master_hm450}
\title{methyl_master_hm450}
\usage{
methyl_master_hm450(
  hm450.input.dir = NULL,
  hm450.output.dir = NULL,
  hm450.sample.sheet.path = NULL,
  hm450.comparison = c("tumor", "normal"),
  hm450.split.by = NULL,
  hm450.reference = "internal",
  hm450.workflow = "B",
  hm450.file.sep = "/",
  hm450.sesame.data.cache = "EPIC",
  hm450.sesame.data.normal = "EPIC.5.normal",
  hm450.genome.version = "hg38",
  hm450.save.seg = FALSE,
  ...
)
}
\arguments{
\item{hm450.input.dir}{The input .idat directory}

\item{hm450.output.dir}{The output directory for the HM450 routine}

\item{hm450.comparison}{The two-element MethylMaster comparison vector
to splut up the HM450 analysis by}

\item{hm450.reference}{The reference to use for the HM450 routine}

\item{hm450.workflow}{The specific HM450 workflow to use: "A", "B", or "C"}

\item{hm450.file.sep}{The system-specific file separator to use}

\item{hm450.sesame.data.cache}{The sesame data control data cache: if using
a sesame dataset as the reference}

\item{hm450.sesame.data.normal}{The sesame data normal data set: if using a
sesame dataset as the reference}

\item{hm450.genome.version}{The sesame ref version (e.g. hg38) if using
a sesame dataset as the reference}

\item{hm450.save.seg}{Whether to save the HM450 routine segmentatin results}

\item{...}{Additional parameters to pass to methyl_master_hm450}

\item{hm450.sample.sheet}{The MethylMaster sample sheet path}
}
\value{
HM450 routine segmentation CNV results stored in a list
}
\description{
MethylMaster version of HM450 analysis
}
