% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_olaps_and_visualize.R
\name{methyl_master_olaps_and_visualize}
\alias{methyl_master_olaps_and_visualize}
\title{methyl_master_olaps_and_visualize}
\usage{
methyl_master_olaps_and_visualize(
  ov.seg = NULL,
  ov.name = NULL,
  ov.output.dir = getwd(),
  ov.routine = NULL,
  ov.split.field = NULL,
  ov.keep.extra.columns = TRUE,
  ov.overlap.density = 0.1,
  ov.estimate.recurrence = FALSE,
  ov.simplify.reduce = weightedmean,
  ov.less.stringent.ra.setting = FALSE,
  ov.pvalue = 0.05,
  ov.plot.individual = FALSE,
  ...
)
}
\value{
#seg.out
}
\description{
perform overlaps analysis and visualization
##Michael Mariani Dartmouth College 2021
}
