% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_plot_individual.R
\name{methyl_master_plot_individual}
\alias{methyl_master_plot_individual}
\title{methyl_master_plot_individual}
\usage{
methyl_master_plot_individual(
  pi.seg = NULL,
  pi.output.dir = NULL,
  pi.name = NULL,
  ...
)
}
\arguments{
\item{pi.seg}{The input sesame segmentation objects}

\item{pi.output.dir}{The output directory}

\item{pi.name}{The names to use}

\item{...}{Additional parameters passed to methyl_master_plot_individual}
}
\value{
Individual plots
}
\description{
Function to output signal inensity plots for individual samples
}
