% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_qreduceassay.R
\name{methyl_master_qreduceassay}
\alias{methyl_master_qreduceassay}
\title{methyl_master_qreduceassay}
\usage{
methyl_master_qreduceassay(
  x,
  query,
  simplifyReduce,
  i = 1,
  withDimnames = TRUE,
  background = NA_integer_
)
}
\arguments{
\item{x}{The x parameter}

\item{query}{The query parameter}

\item{simplifyReduce}{The simplifyReduce parameter}

\item{i}{The i parameter}

\item{withDimnames}{The withDimnames parameter}

\item{background}{the background parameter}
}
\value{
A reduced ReaggedExperiment assay object
}
\description{
MethylMaster version of the qreduceAssay function
Originally from the RaggedExperiment Package by Martin Morgan and
Marcel Ramos 2021.
}
