% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_sesame.R
\name{methyl_master_sesame}
\alias{methyl_master_sesame}
\title{##methyl_master_Sesame}
\usage{
methyl_master_sesame(
  sesame.idat.files.dir = NULL,
  sesame.output.dir = NULL,
  sesame.ref = NULL,
  sesame.sample.sheet.path = NULL,
  sesame.comparison = NULL,
  sesame.file.sep = NULL,
  sesame.data.cache = "EPIC",
  sesame.data.normal = "EPIC.5.normal",
  sesame.ref.version = "hg38",
  sesame.reference = "internal",
  sesame.split.by = NULL,
  sesame.save.seg = FALSE,
  ...
)
}
\value{
A seg object for downstream analysis
}
\description{
MethylMasteR run SeSAMe function

First part:: Very similar to PROCESS SESAME
command to prepare for segementation
this is included so that the total
time and mem for running sesame
can be recorded
}
