% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_sesame.R
\name{methyl_master_sesame}
\alias{methyl_master_sesame}
\title{##Methyl MasteR Sesame}
\usage{
methyl_master_sesame(
  sesame.idat.files.dir = getwd(),
  sesame.samples = NULL,
  sesame.platform = "HM450",
  sesame.data.cache = "EPIC",
  sesame.data.normal = "EPIC.5.normal",
  sesame.ref.version = "hg38",
  sesame.calc.betas = FALSE,
  sesame.infer.sex.karyotypes = FALSE,
  sesamle.ref = NULL,
  sesame.output.dir,
  sesame.reference = "internal",
  sesame.reference.names = c("tumor", "normal"),
  split.by = NULL,
  ...
)
}
\description{
MethylMasteR SeSAMe run

First part:: Very similar to PROCESS SESAME
command to prepare for segementation
this is included so that the total
time and mem for running sesame
can be recorded
}
