% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{read.metharray.mm}
\alias{read.metharray.mm}
\title{Modified read.metharray from minfi}
\usage{
read.metharray.mm(basenames, extended = FALSE, verbose = FALSE, force = FALSE)
}
\value{
methylation array data
}
\description{
Modified read.metharray from minfi
}
