% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{read.metharray.sheet.mm}
\alias{read.metharray.sheet.mm}
\title{Modified read.metharray.sheet from minfi}
\usage{
read.metharray.sheet.mm(
  base,
  pattern = "csv$",
  ignore.case = TRUE,
  recursive = TRUE,
  single.file = TRUE,
  single.file.path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
formatted csv as df
}
\description{
Modified read.metharray.sheet from minfi
}
