% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_epicopy_helper_functions.R
\name{readSheet.mm}
\alias{readSheet.mm}
\title{Modified readSheet from minfi}
\usage{
readSheet.mm(file, recursive = TRUE)
}
\value{
formatted csv as df
}
\description{
Modified readSheet from minfi
}
