% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_plot_individual.R
\name{visualizeSegments.mm}
\alias{visualizeSegments.mm}
\title{visualizeSegments.mm}
\usage{
visualizeSegments.mm(seg, to.plot = NULL)
}
\value{
#A ggplot
}
\description{
modified sesame::visualizeSegments function
Michael Mariani PhD Dartmouth College 2021, originally by SeSAMe team
}
