% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{weightedmean.epicopy}
\alias{weightedmean.epicopy}
\title{weightedmean.epicopy}
\usage{
weightedmean.epicopy(scores, ranges, qranges)
}
\arguments{
\item{scores}{input scores}

\item{ranges}{input Granges}

\item{qranges}{input Qranges}
}
\value{
epicopy-style weighted mean values
}
\description{
Epicopy-style weighted mean function
Original function from Epicopy by Cho et al., 2019
}
