% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_master_helper_functions.R
\name{weightedmean}
\alias{weightedmean}
\title{Weighted-mean function}
\usage{
weightedmean(scores, ranges, qranges)
}
\arguments{
\item{scores}{input scores}

\item{ranges}{input Granges}

\item{qranges}{input Qranges}
}
\value{
weighted mean values
}
\description{
Weighted-mean function
}
